workflowN/requirements.txt
celery
Cython
Django==1.11
django-closuretree
django-debug-toolbar
django-jinja
django-mptt
django-oauth-toolkit==0.11.0
django-python3-ldap==0.11.0
django-redis
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
drf-pdf
elasticsearch-dsl=5.2
fpdf
ldap
Pillow
psycopg2
pylibmc
PyPDF2
python-dateutil
pytz
redis
reportlab
requests
WeasyPrint
xlrd
xlwt
xmltodict
django-celery-email
django-celery-beat
django-celery-results
openpyxl==2.4.9